refactoring the code into mutliple files
